Fixed showing/removing prefs
